requirements.txt
streamlit==1.28.2
torch==2.1.1       
opencv-python==4.7.0.72
ultralytics==8.1.14
numpy==1.24.2      
Pillow==10.4.0     
